src/requirements.txt
tabulate
nibabel
torchio
torchinfo
thop
monai
batchgenerators
timm
einops
# mamba_ssm
triton